# backend/contracts/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "DuplicateProbeRequest": {
        "properties": {
          "identifierValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identifiervalue"
          },
          "partyId": {
            "title": "Partyid",
            "type": "string"
          }
        },
        "required": [
          "partyId"
        ],
        "title": "DuplicateProbeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterPartyReadSchema": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "partyKind": {
            "enum": [
              "natural-person",
              "company",
              "partnership",
              "trust",
              "statutory-body",
              "other"
            ],
            "title": "Partykind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "partyKind"
        ],
        "title": "MatterPartyReadSchema",
        "type": "object"
      },
...
      "PagedResponse_PartyListItemSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PartyListItemSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageSchema"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PagedResponse[PartyListItemSchema]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Draftly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/matters/{matter_id}/parties": {
      "get": {
        "operationId": "list_matter_parties_api_v1_matters__matter_id__parties_get",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterPartyReadSchema"
                  },
                  "title": "Response List Matter Parties Api V1 Matters  Matter Id  Parties Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Matter Parties",
        "tags": [
          "parties"
        ]
      }
    },
...
    "/api/v1/parties": {
      "get": {
        "operationId": "list_parties_api_v1_parties_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_PartyListItemSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Parties",
        "tags": [
          "parties"
        ]
      },
      "post": {
        "operationId": "create_party_api_v1_parties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Party",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/duplicate-probe": {
      "post": {
        "description": "Surfaces candidates. Nothing here merges anything \u2014 see POST /parties/merges.",
        "operationId": "duplicate_probe_api_v1_parties_duplicate_probe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateProbeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Duplicate Probe",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/merges": {
      "post": {
        "description": "Administrator-only, pointer-based, never a delete (party-service.md \u00a75).",
        "operationId": "merge_parties_api_v1_parties_merges_post",
        "parameters": [
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergePartiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merge Parties",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/{party_id}": {
      "get": {
        "operationId": "get_party_api_v1_parties__party_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Party",
        "tags": [
          "parties"
        ]
      },
      "patch": {
        "operationId": "patch_party_api_v1_parties__party_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPartyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Party",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/{party_id}/beneficial-owners": {
      "post": {
        "operationId": "record_beneficial_owner_api_v1_parties__party_id__beneficial_owners_post",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordBeneficialOwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficialOwnerReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Beneficial Owner",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/{party_id}/cdd": {
      "post": {
        "operationId": "record_cdd_api_v1_parties__party_id__cdd_post",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordCddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CddAssessmentReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Cdd",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/{party_id}/identity-evidence": {
      "get": {
        "operationId": "list_identity_evidence_api_v1_parties__party_id__identity_evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IdentityEvidenceReadSchema"
                  },
                  "title": "Response List Identity Evidence Api V1 Parties  Party Id  Identity Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Identity Evidence",
        "tags": [
          "parties"
        ]
      },
      "post": {
        "operationId": "record_identity_evidence_api_v1_parties__party_id__identity_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordIdentityEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityEvidenceReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Identity Evidence",
        "tags": [
          "parties"
        ]
      }
    },
    "/api/v1/parties/{party_id}/identity-evidence/{evidence_id}/identifier": {
      "get": {
        "operationId": "read_identity_value_api_v1_parties__party_id__identity_evidence__evidence_id__identifier_get",
        "parameters": [
          {
            "in": "path",
            "name": "party_id",
            "required": true,
            "schema": {
              "title": "Party Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "purpose",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Purpose",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityValueReadSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Identity Value",
        "tags": [
          "parties"
        ]
      }
    },
...
  }
}